refactor duplicate code
